fix different gua cache conflict
